refactor: DisplayPosition 타입 분리
[#695]
