+++ frontend/src/types/map.ts
@@ -1,0 +1,7 @@
+export interface DisplayPosition {
+  longitude: number;
+  latitude: number;
+  longitudeDelta: number;
+  latitudeDelta: number;
+  zoom: number;
+}
